+++ src/storaged/rbs_sclient.h
@@ -80,7 +80,7 @@ int rbs_get_rb_entry_cnts(rb_entry_t * rbs_restore_one_rb_entry,
  * @return: 0 on success -1 otherwise (errno is set)
  */
 int rbs_get_rb_entry_list(mclient_t * mclt, cid_t cid, sid_t sid,
-        sid_t rebuild_sid, uint8_t * spare, uint8_t * layout,
+        sid_t rebuild_sid, uint8_t * spare, uint32_t * layout,
         sid_t dist_set[ROZOFS_SAFE_MAX], uint64_t * cookie,
         bins_file_rebuild_t ** children, uint8_t * eof);
