+++ src/storaged/storage.c
@@ -613,7 +613,7 @@ out:
 }
 
 int storage_list_bins_files_to_rebuild(storage_t * st, sid_t sid,
-        uint8_t * layout, sid_t *dist_set, uint8_t * spare, uint64_t * cookie,
+        uint32_t * layout, sid_t *dist_set, uint8_t * spare, uint64_t * cookie,
         bins_file_rebuild_t ** children, uint8_t * eof) {
 
     int status = -1;
